Project_3/Proj_3/requirements.txt
asgiref==3.2.7
asn1crypto==0.24.0
attrs==19.3.0
backcall==0.1.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
django-filter==2.2.0
django-filters==0.2.1
dj-database-url==0.5.0
Django==3.0.6
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
entrypoints==0.3
Flask==0.12.2
gunicorn==20.0.4
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
imageio==2.8.0
importlib-metadata==1.6.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==0.24
jedi==0.16.0
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.5
neotime==1.7.4
netifaces==0.10.4
networkx==2.4
nltk==3.4.5
notebook==6.0.3
numpy==1.18.3
oauth==1.0.1
olefile==0.45.1
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
psycopg2-binary==2.8.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.0.0
ptyprocess==0.6.0
py2neo==4.3.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pyinotify==0.9.6
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyRFC3339==1.0
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==3.12
pyzmq==19.0.0
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
SecretStorage==2.3.1
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.13.2
six==1.14.0
sklearn==0.0
soupsieve==2.0
sqlparse==0.3.1
statsmodels==0.11.1
systemd-python==234
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.24.3
wadllib==1.3.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==5.1.0
widgetsnbextension==3.5.1
zipp==3.1.0
zope.interface==4.3.2
